feat - Organized inference folder
