--- src/lane-detection/inference/lane-detection/hailort.log
@@ -1,43 +1,0 @@
-[2026-04-30 12:59:20.477] [99630] [HailoRT] [info] [vdevice.cpp:535] [create] Creating vdevice with params: device_count: 1, scheduling_algorithm: NONE, multi_process_service: false
-[2026-04-30 12:59:20.489] [99630] [HailoRT] [info] [device.cpp:51] [Device] OS Version: Linux 6.6.63-v8-16k #1 SMP PREEMPT Mon Nov  3 10:14:37 UTC 2025 aarch64
-[2026-04-30 12:59:20.490] [99630] [HailoRT] [info] [control.cpp:113] [control__parse_identify_results] firmware_version is: 4.23.0
-[2026-04-30 12:59:20.490] [99630] [HailoRT] [info] [vdevice.cpp:682] [create] VDevice Infos: 0000:01:00.0
-[2026-04-30 12:59:20.497] [99630] [HailoRT] [info] [internal_buffer_manager.cpp:75] [print_execution_results] Default Internal buffer planner failed to meet requirements
-[2026-04-30 12:59:20.497] [99630] [HailoRT] [info] [internal_buffer_manager.cpp:86] [print_execution_results] Default Internal buffer planner executed successfully
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [device_internal.cpp:57] [configure] Configuring HEF took 42.616653 milliseconds
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [vdevice.cpp:780] [configure] Configuring HEF on VDevice took 42.869874 milliseconds
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [core_op.cpp:156] [activate] Activating tusimple_res34_cut_v2 took 0.320332 milliseconds. Note that the function is asynchronous and thus the network is not fully activated yet.
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [core_op.cpp:156] [activate] Activating tusimple_res34_cut_v2 took 0.332184 milliseconds. Note that the function is asynchronous and thus the network is not fully activated yet.
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [edge_elements.cpp:52] [create] Created (HwWriteEl13tusimple_res34_cut_v2/input_layer1 | hw_frame_size: 768000)
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [vstream.cpp:754] [InputVStreamImpl] Creating tusimple_res34_cut_v2/input_layer1...
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [vstream_builder.cpp:102] [create_inputs] Input pipeline 'tusimple_res34_cut_v2/input_layer1': (HwWriteEl13tusimple_res34_cut_v2/input_layer1 | hw_frame_size: 768000) >> HW
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [edge_elements.cpp:287] [create] Created (HwReadEl12tusimple_res34_cut_v2/conv37 | hw_frame_size: 2000)
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [queue_elements.cpp:800] [create] Created (UserBuffQEl12tusimple_res34_cut_v2/conv37 | timeout: 10s)
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [vstream.cpp:1114] [OutputVStreamImpl] Creating tusimple_res34_cut_v2/conv37...
-[2026-04-30 12:59:20.533] [99630] [HailoRT] [info] [vstream_builder.cpp:220] [create_outputs] Output pipeline 'tusimple_res34_cut_v2/conv37': HW >> (HwReadEl12tusimple_res34_cut_v2/conv37 | hw_frame_size: 2000) >> (UserBuffQEl12tusimple_res34_cut_v2/conv37 | timeout: 10s)
-[2026-04-30 12:59:22.357] [99664] [HailoRT] [info] [queue_elements.cpp:855] [run_in_thread] Shutdown event was signaled in dequeue of UserBuffQEl12tusimple_res34_cut_v2/conv37!
-[2026-04-30 12:59:22.357] [99630] [HailoRT] [info] [queue_elements.cpp:46] [~BaseQueueElement] Queue element UserBuffQEl12tusimple_res34_cut_v2/conv37 has 0 frames in his Queue on destruction
-[2026-04-30 12:59:22.358] [99630] [HailoRT] [info] [core_op.cpp:186] [deactivate] Deactivating took 0.425109 ms
-[2026-04-30 12:59:22.358] [99630] [HailoRT] [info] [core_op.cpp:186] [deactivate] Deactivating took 0.441887 ms
-[2026-04-30 14:31:51.831] [78290] [HailoRT] [info] [vdevice.cpp:535] [create] Creating vdevice with params: device_count: 1, scheduling_algorithm: NONE, multi_process_service: false
-[2026-04-30 14:31:51.844] [78290] [HailoRT] [info] [device.cpp:51] [Device] OS Version: Linux 6.6.63-v8-16k #1 SMP PREEMPT Mon Nov  3 10:14:37 UTC 2025 aarch64
-[2026-04-30 14:31:51.844] [78290] [HailoRT] [info] [control.cpp:113] [control__parse_identify_results] firmware_version is: 4.23.0
-[2026-04-30 14:31:51.844] [78290] [HailoRT] [info] [vdevice.cpp:682] [create] VDevice Infos: 0000:01:00.0
-[2026-04-30 14:31:51.850] [78290] [HailoRT] [info] [internal_buffer_manager.cpp:75] [print_execution_results] Default Internal buffer planner failed to meet requirements
-[2026-04-30 14:31:51.850] [78290] [HailoRT] [info] [internal_buffer_manager.cpp:86] [print_execution_results] Default Internal buffer planner executed successfully
-[2026-04-30 14:31:51.886] [78290] [HailoRT] [info] [device_internal.cpp:57] [configure] Configuring HEF took 41.980124 milliseconds
-[2026-04-30 14:31:51.886] [78290] [HailoRT] [info] [vdevice.cpp:780] [configure] Configuring HEF on VDevice took 42.23179 milliseconds
-[2026-04-30 14:31:51.887] [78290] [HailoRT] [info] [core_op.cpp:156] [activate] Activating tusimple_res34_cut_v2 took 0.404573 milliseconds. Note that the function is asynchronous and thus the network is not fully activated yet.
-[2026-04-30 14:31:51.887] [78290] [HailoRT] [info] [core_op.cpp:156] [activate] Activating tusimple_res34_cut_v2 took 0.418018 milliseconds. Note that the function is asynchronous and thus the network is not fully activated yet.
-[2026-04-30 14:31:51.887] [78290] [HailoRT] [info] [edge_elements.cpp:52] [create] Created (HwWriteEl13tusimple_res34_cut_v2/input_layer1 | hw_frame_size: 768000)
-[2026-04-30 14:31:51.887] [78290] [HailoRT] [info] [vstream.cpp:754] [InputVStreamImpl] Creating tusimple_res34_cut_v2/input_layer1...
-[2026-04-30 14:31:51.887] [78290] [HailoRT] [info] [vstream_builder.cpp:102] [create_inputs] Input pipeline 'tusimple_res34_cut_v2/input_layer1': (HwWriteEl13tusimple_res34_cut_v2/input_layer1 | hw_frame_size: 768000) >> HW
-[2026-04-30 14:31:51.887] [78290] [HailoRT] [info] [edge_elements.cpp:287] [create] Created (HwReadEl12tusimple_res34_cut_v2/conv37 | hw_frame_size: 2000)
-[2026-04-30 14:31:51.887] [78290] [HailoRT] [info] [queue_elements.cpp:800] [create] Created (UserBuffQEl12tusimple_res34_cut_v2/conv37 | timeout: 10s)
-[2026-04-30 14:31:51.887] [78290] [HailoRT] [info] [vstream.cpp:1114] [OutputVStreamImpl] Creating tusimple_res34_cut_v2/conv37...
-[2026-04-30 14:31:51.887] [78290] [HailoRT] [info] [vstream_builder.cpp:220] [create_outputs] Output pipeline 'tusimple_res34_cut_v2/conv37': HW >> (HwReadEl12tusimple_res34_cut_v2/conv37 | hw_frame_size: 2000) >> (UserBuffQEl12tusimple_res34_cut_v2/conv37 | timeout: 10s)
-[2026-04-30 14:32:04.253] [78326] [HailoRT] [info] [queue_elements.cpp:855] [run_in_thread] Shutdown event was signaled in dequeue of UserBuffQEl12tusimple_res34_cut_v2/conv37!
-[2026-04-30 14:32:04.253] [78326] [HailoRT] [info] [queue_elements.cpp:68] [operator()] Thread in element UserBuffQEl12tusimple_res34_cut_v2/conv37 is not running anymore, exiting..
-[2026-04-30 14:32:04.253] [78290] [HailoRT] [info] [queue_elements.cpp:46] [~BaseQueueElement] Queue element UserBuffQEl12tusimple_res34_cut_v2/conv37 has 0 frames in his Queue on destruction
-[2026-04-30 14:32:04.254] [78290] [HailoRT] [info] [core_op.cpp:186] [deactivate] Deactivating took 0.59648 ms
-[2026-04-30 14:32:04.254] [78290] [HailoRT] [info] [core_op.cpp:186] [deactivate] Deactivating took 0.611332 ms
