+++ src/lane-detection/inference/lane-detection/rc_inference-with-kuksa-v3.py
@@ -1,0 +1,482 @@
+import json
+import numpy as np
+import cv2
+import threading
+import queue
+import time
+import sys
+from hailo_platform import (HEF, VDevice, HailoStreamInterface,
+                             InferVStreams, ConfigureParams,
+                             InputVStreamParams, OutputVStreamParams,
+                             FormatType)
+import grpc
+from kuksa.val.v2 import val_pb2, val_pb2_grpc, types_pb2
+
+import sys, os
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+from config import (
+    HEF_PATH, WEIGHTS_PATH, OUTPUT_PATH,
+    FULL_W, FULL_H, CAM_FPS,
+    TRAIN_WIDTH, TRAIN_HEIGHT, CROP_RATIO,
+    INPUT_STREAM, OUTPUT_LAYER,
+    EXIST_THRESHOLD, TEMPORAL_HISTORY, TEMPORAL_MIN_HITS,
+    KUKSA_HOST, KUKSA_PORT, KUKSA_CA_CERT, KUKSA_TOKEN,
+    LKA_DEVIATION_PATH, LKA_STATUS_PATH, CAMERA_OFFSET_CM,
+)
+
+LKA_WAYPOINTS_PATH  = "Vehicle.ADAS.LaneKeepAssist.LaneWaypoints"
+LKA_POLYNOMIAL_PATH = "Vehicle.ADAS.LaneKeepAssist.LanePolynomial"
+
+from postprocess import (
+    load_weights, postprocess, decode_lanes,
+    calc_lateral_deviation, TemporalLaneSmoother, smooth_lanes,
+)
+from visualization import draw_lanes, draw_drivable_area, draw_overlay
+from pathlib import Path
+
+
+# ── Calibração da câmara ──────────────────────────────────────────────────────
+CALIB_DIR_FRONT = Path("/data/seame-configs/camera")
+CALIB_DIR_REAR  = Path("/data/seame-configs/camera_rear")
+INTRINSIC_DIR   = Path("/data/seame-configs/camera")  # shared — same lens
+
+
+class Calibration:
+    """Loads intrinsic + extrinsic calibration for undistortion and BEV."""
+
+    def __init__(self, is_rear=False):
+        self.enabled = False
+        self._undist_map1 = None
+        self._undist_map2 = None
+        self.H_img2world = None
+
+        calib_dir = CALIB_DIR_REAR if is_rear else CALIB_DIR_FRONT
+        cam_label = "REAR" if is_rear else "FRONT"
+
+        files = {
+            "camera_matrix.npy": INTRINSIC_DIR / "camera_matrix.npy",
+            "dist_coeffs.npy": INTRINSIC_DIR / "dist_coeffs.npy",
+            "homography_img2world.npy": calib_dir / "homography_img2world.npy",
+        }
+        missing = [n for n, p in files.items() if not p.exists()]
+        if missing:
+            print(f"[CALIB] [{cam_label}] Ficheiros em falta: {', '.join(missing)}")
+            print(f"[CALIB] A correr SEM calibração")
+            return
+
+        camera_matrix = np.load(files["camera_matrix.npy"])
+        dist_coeffs = np.load(files["dist_coeffs.npy"])
+        self.H_img2world = np.load(files["homography_img2world.npy"])
+
+        new_matrix, _ = cv2.getOptimalNewCameraMatrix(
+            camera_matrix, dist_coeffs,
+            (FULL_W, FULL_H), 0, (FULL_W, FULL_H)
+        )
+        self._undist_map1, self._undist_map2 = cv2.initUndistortRectifyMap(
+            camera_matrix, dist_coeffs, None, new_matrix,
+            (FULL_W, FULL_H), cv2.CV_16SC2
+        )
+        self.enabled = True
+        print(f"[CALIB] [{cam_label}] OK — intrinsic (shared) + extrinsic ({calib_dir.name}/) loaded")
+
+    def undistort(self, frame):
+        if not self.enabled:
+            return frame
+        return cv2.remap(frame, self._undist_map1, self._undist_map2,
+                         cv2.INTER_LINEAR)
+
+    def pixels_to_world(self, points_px):
+        """Convert list of (x,y) pixel tuples to world coords (cm)."""
+        if not self.enabled or self.H_img2world is None:
+            return None
+        pts = np.array(points_px, dtype=np.float32).reshape(-1, 1, 2)
+        return cv2.perspectiveTransform(pts, self.H_img2world).reshape(-1, 2)
+
+
+def calc_lateral_deviation_cm(lanes, calib):
+    """
+    Calculate lateral deviation in cm using homography.
+    Falls back to pixel-based calc if calibration is unavailable.
+    """
+    from postprocess import calc_lateral_deviation as _fallback
+
+    def bottom_points(pts, n=5):
+        if len(pts) < 3:
+            return None
+        sorted_pts = sorted(pts, key=lambda p: p[1], reverse=True)
+        return sorted_pts[:n]
+
+    left_pts = bottom_points(lanes[1])
+    right_pts = bottom_points(lanes[2])
+
+    has_left = left_pts is not None
+    has_right = right_pts is not None
+
+    if not has_left and not has_right:
+        return None, "none"
+
+    if has_left:
+        left_world = calib.pixels_to_world(left_pts)
+        left_x = float(np.mean(left_world[:, 0]))
+    if has_right:
+        right_world = calib.pixels_to_world(right_pts)
+        right_x = float(np.mean(right_world[:, 0]))
+
+    if has_left and has_right:
+        lane_center = (left_x + right_x) / 2.0
+        deviation = -lane_center - CAMERA_OFFSET_CM
+        return deviation, "both"
+
+    if has_left:
+        deviation = -(left_x + 15.0) - CAMERA_OFFSET_CM
+        return deviation, "left"
+
+    deviation = -(right_x - 15.0) - CAMERA_OFFSET_CM
+    return deviation, "right"
+
+
+# ── KUKSA ──────────────────────────────────────────────────────────────────────
+_kuksa_queue = queue.Queue(maxsize=1)
+
+
+def _kuksa_worker(q):
+    token    = open(KUKSA_TOKEN).read().strip()
+    ca_certs = open(KUKSA_CA_CERT, "rb").read()
+    creds    = grpc.ssl_channel_credentials(root_certificates=ca_certs)
+    channel  = grpc.secure_channel(f"{KUKSA_HOST}:{KUKSA_PORT}", creds)
+    stub     = val_pb2_grpc.VALStub(channel)
+    metadata = [("authorization", f"Bearer {token}")]
+    print("[KUKSA] Thread iniciada (gRPC v2)")
+    while True:
+        deviation, status, waypoints_json, poly_json = q.get()
+        try:
+            req_dev = val_pb2.PublishValueRequest()
+            req_dev.signal_id.path = LKA_DEVIATION_PATH
+            req_dev.data_point.value.float = float(deviation)
+            stub.PublishValue(req_dev, metadata=metadata)
+
+            req_st = val_pb2.PublishValueRequest()
+            req_st.signal_id.path = LKA_STATUS_PATH
+            req_st.data_point.value.string = status
+            stub.PublishValue(req_st, metadata=metadata)
+
+            req_wp = val_pb2.PublishValueRequest()
+            req_wp.signal_id.path = LKA_WAYPOINTS_PATH
+            req_wp.data_point.value.string = waypoints_json
+            stub.PublishValue(req_wp, metadata=metadata)
+
+            req_poly = val_pb2.PublishValueRequest()
+            req_poly.signal_id.path = LKA_POLYNOMIAL_PATH
+            req_poly.data_point.value.string = poly_json
+            stub.PublishValue(req_poly, metadata=metadata)
+        except Exception as e:
+            print(f"[KUKSA] Erro: {e}")
+
+
+def publish_lane_data(deviation, status, waypoints_json, poly_json):
+    try:
+        _kuksa_queue.put_nowait((
+            deviation if deviation is not None else 0.0,
+            status,
+            waypoints_json,
+            poly_json,
+        ))
+    except queue.Full:
+        pass
+
+
+# ── Writer assíncrono ──────────────────────────────────────────────────────────
+class AsyncVideoWriter:
+    def __init__(self, path, fps, width, height):
+        self._queue  = queue.Queue(maxsize=60)
+        self._writer = cv2.VideoWriter(
+            path, cv2.VideoWriter_fourcc(*"mp4v"), fps, (width, height))
+        if not self._writer.isOpened():
+            raise RuntimeError(f"Não foi possível criar: {path}")
+        self._stop   = threading.Event()
+        self._thread = threading.Thread(target=self._loop, daemon=True)
+        self._thread.start()
+
+    def _loop(self):
+        while not self._stop.is_set() or not self._queue.empty():
+            try:
+                frame = self._queue.get(timeout=0.1)
+                self._writer.write(frame)
+            except Exception:
+                continue
+
+    def write(self, frame):
+        try:
+            self._queue.put_nowait(frame)
+        except queue.Full:
+            pass
+
+    def release(self):
+        self._stop.set()
+        self._thread.join(timeout=10)
+        self._writer.release()
+
+
+# ── Pré-processamento ──────────────────────────────────────────────────────────
+def preprocess(frame):
+    """
+    BGR FULL_W×FULL_H → UINT8 (320, 800, 3)
+    Resize 800×400 → crop inferior 800×320 → RGB
+    """
+    resize_h = int(TRAIN_HEIGHT / CROP_RATIO)  # 400
+    img = cv2.resize(frame, (TRAIN_WIDTH, resize_h))
+    top = resize_h - TRAIN_HEIGHT               # 80
+    img = img[top:, :]
+    img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
+    return img.astype(np.uint8)
+
+
+# ── Lane trajectory ────────────────────────────────────────────────────────────
+def compute_lane_trajectory(lanes, calib):
+    """
+    Build waypoints and fit a 2nd-degree polynomial from the detected lanes.
+
+    Waypoints are the centre-line in world coordinates (cm): [[x, y], ...],
+    sorted by increasing y (near → far).  Returns (waypoints_json, poly_json).
+    poly_json is "[a, b, c]" such that x_lateral = a·y² + b·y + c.
+    Either string is "" when there is insufficient data.
+    """
+    left_pts  = lanes[1]   # [(x_px, y_px), ...]
+    right_pts = lanes[2]   # [(x_px, y_px), ...]
+
+    has_left  = len(left_pts)  >= 3
+    has_right = len(right_pts) >= 3
+
+    if not has_left and not has_right:
+        return "", ""
+
+    # Build centre-line points in pixel space first, then project to world.
+    if has_left and has_right:
+        # Match left/right by closest y, average x.
+        left_arr  = np.array(sorted(left_pts,  key=lambda p: p[1]), dtype=np.float32)
+        right_arr = np.array(sorted(right_pts, key=lambda p: p[1]), dtype=np.float32)
+        # Sample at left's y values; interpolate right x for each.
+        right_x_interp = np.interp(left_arr[:, 1], right_arr[:, 1], right_arr[:, 0])
+        center_px = np.stack(
+            [(left_arr[:, 0] + right_x_interp) / 2.0, left_arr[:, 1]], axis=1
+        )
+    elif has_left:
+        center_px = np.array(sorted(left_pts, key=lambda p: p[1]), dtype=np.float32)
+    else:
+        center_px = np.array(sorted(right_pts, key=lambda p: p[1]), dtype=np.float32)
+
+    # Project to world (cm) if calibration available.
+    if calib.enabled:
+        pts_world = calib.pixels_to_world(center_px.tolist())
+        if pts_world is None:
+            return "", ""
+        # Sort by distance (y in world = forward distance)
+        pts_world = pts_world[np.argsort(pts_world[:, 1])]
+        waypoints = pts_world.tolist()
+    else:
+        # Fall back to normalised pixel coords (no cm units)
+        pts_world = center_px[np.argsort(center_px[:, 1])]
+        waypoints = pts_world.tolist()
+
+    waypoints_json = json.dumps([[round(p[0], 2), round(p[1], 2)] for p in waypoints])
+
+    # Fit x = a·y² + b·y + c (need ≥ 3 points)
+    if len(waypoints) >= 3:
+        ys = np.array([p[1] for p in waypoints])
+        xs = np.array([p[0] for p in waypoints])
+        try:
+            coeffs = np.polyfit(ys, xs, 2)
+            poly_json = json.dumps([round(float(c), 6) for c in coeffs])
+        except np.linalg.LinAlgError:
+            poly_json = ""
+    else:
+        poly_json = ""
+
+    return waypoints_json, poly_json
+
+
+# ── Pipeline principal ─────────────────────────────────────────────────────────
+def run(frame_queue, duration_seconds=60, save_video=False, is_rear=False):
+    global CAMERA_OFFSET_CM
+    threading.Thread(target=_kuksa_worker, args=(_kuksa_queue,), daemon=True).start()
+
+    # Load rear config overrides if needed
+    if is_rear:
+        try:
+            from config_rear import CAMERA_OFFSET_CM as _rear_offset
+            CAMERA_OFFSET_CM = _rear_offset
+            print(f"[CONFIG] Rear camera offset: {CAMERA_OFFSET_CM} cm")
+        except ImportError:
+            print(f"[CONFIG] config_rear.py not found — using front offset ({CAMERA_OFFSET_CM} cm)")
+
+    W1, b1, W2, b2, keep_rows = load_weights(WEIGHTS_PATH)
+    calib = Calibration(is_rear=is_rear)
+
+    print(f"\nFrame câmara:      {FULL_W}×{FULL_H} (via CameraBroker)")
+    print(f"Modelo recebe:     {TRAIN_WIDTH}×{TRAIN_HEIGHT} (crop_ratio={CROP_RATIO})")
+    print(f"HEF cortado:       /pool/Conv → conv37 (10,25,8) UINT8")
+    print(f"Post-processing:   CPU (dequant + FC1 + ReLU + FC2)")
+    print(f"EXIST_THRESHOLD:   {EXIST_THRESHOLD}")
+    print(f"Temporal smoother: history={TEMPORAL_HISTORY} min_hits={TEMPORAL_MIN_HITS}")
+
+    async_writer = None
+    if save_video:
+        async_writer = AsyncVideoWriter(OUTPUT_PATH, CAM_FPS, FULL_W, FULL_H)
+        print(f"A gravar: {OUTPUT_PATH}")
+
+    hef = HEF(HEF_PATH)
+    with VDevice() as target:
+        configure_params = ConfigureParams.create_from_hef(
+            hef, interface=HailoStreamInterface.PCIe)
+        network_groups = target.configure(hef, configure_params)
+        network_group  = network_groups[0]
+
+        input_params  = InputVStreamParams.make(network_group, format_type=FormatType.UINT8)
+        output_params = OutputVStreamParams.make(network_group, format_type=FormatType.UINT8)
+
+        with network_group.activate():
+            with InferVStreams(network_group, input_params, output_params) as pipeline:
+                print(f"\nA correr durante {duration_seconds}s — Ctrl+C para parar\n")
+                print(f"{'Frame':<7} {'Pre':>6} {'Hailo':>7} {'Post':>7} "
+                      f"{'Coords':>8} {'Total':>7} {'Lanes':>6} {'Dev':>8} {'Status'}")
+                print("─" * 78)
+
+                frame_idx         = 0
+                t_start           = time.time()
+                times_all         = []
+                fps_acc           = []
+                temporal_smoother = TemporalLaneSmoother()
+                prev_status       = None
+                single_lane_mode  = False
+
+                try:
+                    while (time.time() - t_start) < duration_seconds:
+                        try:
+                            frame = frame_queue.get(timeout=0.1)
+                        except queue.Empty:
+                            continue
+
+                        t0    = time.time()
+                        frame = calib.undistort(frame)
+                        img   = preprocess(frame)
+                        t_pre = (time.time() - t0) * 1000
+
+                        img        = np.ascontiguousarray(img)
+                        t0         = time.time()
+                        input_data = {INPUT_STREAM: img[np.newaxis]}
+                        output     = pipeline.infer(input_data)
+                        t_hailo    = (time.time() - t0) * 1000
+
+                        t0     = time.time()
+                        conv37 = output[OUTPUT_LAYER][0]
+                        loc_row_lanes12, exist_row_lanes12 = \
+                            postprocess(conv37, W1, b1, W2, b2, keep_rows)
+                        lanes  = decode_lanes(loc_row_lanes12, exist_row_lanes12,
+                                              single_lane_mode=single_lane_mode)
+                        t_post = (time.time() - t0) * 1000
+
+                        t0                = time.time()
+                        lanes             = temporal_smoother.update(lanes)
+                        lanes             = smooth_lanes(lanes)
+                        if calib.enabled:
+                            deviation, status = calc_lateral_deviation_cm(lanes, calib)
+                        else:
+                            deviation, status = calc_lateral_deviation(lanes)
+                        waypoints_json, poly_json = compute_lane_trajectory(lanes, calib)
+                        t_coords          = (time.time() - t0) * 1000
+
+                        publish_lane_data(deviation, status, waypoints_json, poly_json)
+
+                        if prev_status == "both" and status in ("left", "right"):
+                            single_lane_mode = True
+                        elif status == "both":
+                            single_lane_mode = False
+                        prev_status = status
+
+                        t_total   = t_pre + t_hailo + t_post + t_coords
+                        num_lanes = sum(1 for l in lanes if l)
+                        times_all.append(t_total)
+
+                        fps_acc.append(1000.0 / (t_total + 1e-9))
+                        if len(fps_acc) > 30:
+                            fps_acc.pop(0)
+                        fps = sum(fps_acc) / len(fps_acc)
+
+                        dev_str = (f"{deviation:+.1f}cm" if calib.enabled and deviation is not None
+                                   else f"{deviation:+.3f}" if deviation is not None
+                                   else "   N/A")
+                        print(f"{frame_idx:<7} "
+                              f"{t_pre:>5.1f}ms "
+                              f"{t_hailo:>6.1f}ms "
+                              f"{t_post:>6.1f}ms "
+                              f"{t_coords:>7.1f}ms "
+                              f"{t_total:>6.1f}ms "
+                              f"{num_lanes:>6} "
+                              f"{dev_str:>8} "
+                              f"{status}")
+
+                        if save_video and async_writer is not None:
+                            frame_out = draw_drivable_area(frame.copy(), lanes)
+                            frame_out = draw_lanes(frame_out, lanes)
+                            frame_out = draw_overlay(
+                                frame_out, fps, frame_idx,
+                                t_pre, t_hailo, t_post, t_coords,
+                                deviation, status)
+                            async_writer.write(frame_out)
+
+                        frame_idx += 1
+
+                except KeyboardInterrupt:
+                    print("\nInterrompido.")
+
+    if async_writer is not None:
+        print("\nA aguardar gravação...")
+        async_writer.release()
+        print(f"Vídeo guardado: {OUTPUT_PATH}")
+
+    elapsed  = time.time() - t_start
+    real_fps = frame_idx / elapsed if elapsed > 0 else 0
+    avg      = sum(times_all) / len(times_all) if times_all else 0
+    sorted_t = sorted(times_all)
+    p50      = sorted_t[int(len(sorted_t) * 0.50)] if sorted_t else 0
+    p95      = sorted_t[int(len(sorted_t) * 0.95)] if sorted_t else 0
+
+    print("─" * 78)
+    print(f"\nResumo ({frame_idx} frames em {elapsed:.1f}s):")
+    print(f"   FPS real:          {real_fps:.1f}")
+    print(f"   FPS pipeline puro: {1000/avg:.1f}" if avg > 0 else "   FPS pipeline puro: N/A")
+    print(f"   Latência média:    {avg:.1f}ms")
+    print(f"   P50:               {p50:.1f}ms")
+    print(f"   P95:               {p95:.1f}ms")
+
+
+if __name__ == "__main__":
+    import argparse
+    sys.path.insert(0, "/opt/seame/adas")
+    from camera_broker import CameraBroker
+
+    p = argparse.ArgumentParser(description="LKA Inference — UFLD + Kuksa")
+    p.add_argument("duration", type=int, nargs="?", default=60,
+                   help="Duração em segundos (default: 60)")
+    p.add_argument("--save",   action="store_true",
+                   help="Gravar vídeo anotado")
+    p.add_argument("--camera", type=int, default=1,
+                   help="Índice da câmara: 1=frente (default), 0=trás")
+    p.add_argument("--flip",   type=int, default=None,
+                   help="Flip: 0=vertical, 1=horizontal, -1=180° (câmara traseira)")
+    args = p.parse_args()
+
+    cam_label = "TRASEIRA (flip=-1)" if args.camera == 0 else "FRONTAL"
+    print(f"SEAME Lane Detection | TuSimple cut v2 | {TRAIN_WIDTH}×{TRAIN_HEIGHT}")
+    print(f"Câmara: {cam_label} (index={args.camera})")
+    print(f"HEF cortado em /pool/Conv | Post-processing CPU\n")
+
+    broker = CameraBroker(width=FULL_W, height=FULL_H, fps=CAM_FPS,
+                          camera=args.camera, flip=args.flip)
+    q = broker.register("lane_detection")
+    broker.start()
+
+    run(q, duration_seconds=args.duration, save_video=args.save,
+        is_rear=(args.camera == 0))
+
+    broker.stop()
